Create a `public-services.new` route
Extracting the public service creation logic into a route allows us to create public services from other routes without having to duplicate the logic and loading templates.
