+++ app/components/shared/bread-crumb.js
@@ -299,6 +299,13 @@ export default class SharedBreadCrumbComponent extends Component {
       ],
     },
     {
+      route: 'public-services.new',
+      crumbs: [
+        { label: 'Producten- en dienstencatalogus', link: 'public-services' },
+        { label: 'Product of dienst toevoegen' },
+      ],
+    },
+    {
       route: 'public-services.details',
       crumbs: [
         { label: 'Producten- en dienstencatalogus', link: 'public-services' },
