+++ app/controllers/public-services/new.js
@@ -1,0 +1,5 @@
+import Controller from '@ember/controller';
+
+export default class PublicServicesNewController extends Controller {
+  queryParams = ['concept'];
+}
