+++ app/router.js
@@ -105,6 +105,7 @@ Router.map(function () {
       { path: '/producten-en-dienstencatalogus' },
       function () {
         this.route('add', { path: '/toevoegen' });
+        this.route('new', { path: '/nieuw' });
         this.route('details', { path: '/:serviceId' }, function () {
           this.route('content', { path: '/inhoud' });
           this.route('properties', { path: '/eigenschappen' });
