+++ app/templates/public-services/new.hbs
@@ -1,0 +1,6 @@
+{{page-title "Nieuw product of dienst aan het toevoegen"}}
+
+<div class="au-o-box">
+  <AuHeading @skin="4">Nieuw product of dienst aan het toevoegen</AuHeading>
+  <AuLoader @padding="small" />
+</div>
